Extends from Abstract CacheEntry. Implement a generic cache entry file for key: value cases.
